requirements.txt
anyio==3.6.2
ariadne==0.17.1
certifi==2022.12.7
click==8.0.4
Flask==2.2.2
Flask-Cors==3.0.10
graphql-core==3.2.3
idna==3.4
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
pip==22.3.1
python-dotenv==0.21.0
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
starlette==0.23.1
typing_extensions==4.4.0
Werkzeug==2.2.2
wheel==0.37.1
